requirements.txt
robotframework
robotframework-browser
playwright
screeninfo
python-dotenv
psutil
Jinja2
flask